Saving version 1.38.0 before deploying to PyPI.

+++ barbell2/npy2png.py
@@ -41,7 +41,7 @@ class Numpy2Png:
             npy_array = self.npy_array_or_file_path
         npy_array = apply_window(npy_array, self.window)
         if self.color_map is not None:
-            npy_array = apply_color_map(npy_array, self.color_map)
+            npy_array = apply_color_map(npy_array, self.color_map, dtype=float)
         fig = plt.figure(figsize=self.png_figure_size)
         ax = fig.add_subplot(1, 1, 1)
         if self.color_map is not None:
